# perf(cli): Improve concurrency when setting up `node_modules` and loading cached npm package info